fix(mcp): isolate Windows schema workers
